+++ src/lib.rs
@@ -165,91 +165,6 @@ impl Drop for Lock<'_> {
     }
 }
 
-/// Payload struct to register or update entries in consul's catalog.
-/// See https://www.consul.io/api-docs/catalog#register-entity for more information.
-#[allow(non_snake_case)]
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct RegisterEntityPayload {
-    /// Optional UUID to assign to the node. This string is required to be 36-characters and UUID formatted.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ID: Option<String>,
-    /// Node ID to register.
-    pub Node: String,
-    /// The address to register.
-    pub Address: String,
-    /// The datacenter to register in, defaults to the agent's datacenter.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub Datacenter: Option<String>,
-    /// Tagged addressed to register with.
-    #[serde(skip_serializing_if = "HashMap::is_empty")]
-    pub TaggedAddresses: HashMap<String, String>,
-    /// KV metadata paris to register with.
-    #[serde(skip_serializing_if = "HashMap::is_empty")]
-    pub NodeMeta: HashMap<String, String>,
-    /// Optional service to register.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub Service: Option<RegisterEntityService>,
-    /// Optional check to register
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub Check: Option<RegisterEntityCheck>,
-    /// Whether to skip updating the nodes information in the registration.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub SkipNodeUpdate: Option<bool>,
-}
-/// The service to register with consul's global catalog.
-/// See https://www.consul.io/api/agent/service for more information.
-#[allow(non_snake_case)]
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct RegisterEntityService {
-    /// ID to register service will, defaults to Service.Service property.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ID: Option<String>,
-    /// The name of the service.
-    pub Service: String,
-    /// Optional tags associated with the service.
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub Tags: Vec<String>,
-    /// Optional map of explicit LAN and WAN addresses for the service.
-    #[serde(skip_serializing_if = "HashMap::is_empty")]
-    pub TaggedAddresses: HashMap<String, String>,
-    /// Optional key value meta associated with the service.
-    #[serde(skip_serializing_if = "HashMap::is_empty")]
-    pub Meta: HashMap<String, String>,
-    /// The port of the service
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub Port: Option<u16>,
-    /// The consul namespace to register the service in.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub Namespace: Option<String>,
-}
-/// Information related to registering a check.
-/// See https://www.consul.io/docs/discovery/checks for more information.
-#[allow(non_snake_case)]
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct RegisterEntityCheck {
-    /// The node to execute the check on.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub Node: Option<String>,
-    /// Optional check id, defaults to the name of the check.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub CheckID: Option<String>,
-    /// The name associated with the check
-    pub Name: String,
-    /// Opaque field encapsulating human-readable text.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub Notes: Option<String>,
-    /// The status of the check. Must be one of 'passing', 'warning', or 'critical'.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub Status: Option<String>,
-    /// ID of the service this check is for. If no ID of a service running on the node is provided,
-    /// the check is treated as a node level check
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ServiceID: Option<String>,
-    /// Details for a TCP or HTTP health check.
-    #[serde(skip_serializing_if = "HashMap::is_empty")]
-    pub Definition: HashMap<String, String>,
-}
-
 #[derive(Debug)]
 /// This struct defines the consul client and allows access to the consul api via method syntax.
 pub struct Consul {
@@ -386,11 +301,7 @@ impl Consul {
     /// - request - the [LockRequest](consul::types::LockRequest)
     /// # Errors:
     /// [ConsulError](consul::ConsulError) describes all possible errors returned by this api.
-    pub async fn get_lock<'a, 'b>(
-        &'b self,
-        request: LockRequest<'a>,
-        value: &[u8],
-    ) -> Result<Lock<'b>> {
+    pub async fn get_lock(&self, request: LockRequest<'_>, value: &[u8]) -> Result<Lock<'_>> {
         let session = self.get_session(request).await?;
         let req = CreateOrUpdateKeyRequest {
             key: request.key,
@@ -450,7 +361,6 @@ impl Consul {
         self.read_key(req).await
     }
 
-    // TODO allow clients to specify per-endpoint timeouts
     /// Registers or updates entries in consul's global catalog.
     /// See https://www.consul.io/api-docs/catalog#register-entity for more information.
     /// # Arguments:
@@ -466,28 +376,34 @@ impl Consul {
         Ok(())
     }
 
-    /// Returns the services currently registered with consul for the datacenter of the agent being queried.
+    /// Returns all services currently registered with consul.
     /// See https://www.consul.io/api-docs/catalog#list-services for more information.
     /// # Arguments:
-    /// - namespace: The namespace of the services to list, defaults to the 'default' namespace.
+    /// - query_opts: The [`QueryOptions`](QueryOptions) to apply for this endpoint.
     /// # Errors:
     /// [ConsulError](consul::ConsulError) describes all possible errors returned by this api.
     pub async fn get_all_registered_service_names(
         &self,
-        namespace: Option<String>,
-    ) -> Result<Vec<String>> {
+        query_opts: Option<QueryOptions>,
+    ) -> Result<ResponseMeta<Vec<String>>> {
         let mut uri = format!("{}/v1/catalog/services", self.config.address);
-        if let Some(ns) = namespace {
-            uri.push_str(&format!("?ns={}", ns));
-        }
-        let request = hyper::Request::builder().method(Method::GET).uri(uri);
-        let (mut response_body, _) = self
-            .execute_request(request, hyper::Body::empty(), Some(Duration::from_secs(5)))
+        let query_opts = query_opts.unwrap_or_default();
+        add_query_option_params(&mut uri, &query_opts, '?');
+
+        let request = hyper::Request::builder()
+            .method(Method::GET)
+            .uri(uri.clone());
+        let (mut response_body, index) = self
+            .execute_request(request, hyper::Body::empty(), query_opts.timeout)
             .await?;
         let bytes = response_body.copy_to_bytes(response_body.remaining());
         let service_tags_by_name = serde_json::from_slice::<HashMap<String, Vec<String>>>(&bytes)
             .map_err(ConsulError::ResponseDeserializationFailed)?;
-        Ok(service_tags_by_name.keys().cloned().collect())
+
+        Ok(ResponseMeta {
+            response: service_tags_by_name.keys().cloned().collect(),
+            index,
+        })
     }
 
     /// returns the nodes providing the service indicated on the path.
@@ -500,31 +416,31 @@ impl Consul {
     pub async fn get_service_nodes(
         &self,
         request: GetServiceNodesRequest<'_>,
-    ) -> Result<GetServiceNodesResponse> {
-        let req = self.build_get_service_nodes_req(request);
-        let (mut response_body, _index) = self
-            .execute_request(
-                req,
-                hyper::Body::empty(),
-                Some(std::time::Duration::from_secs(5)),
-            )
+        query_opts: Option<QueryOptions>,
+    ) -> Result<ResponseMeta<GetServiceNodesResponse>> {
+        let query_opts = query_opts.unwrap_or_default();
+        let req = self.build_get_service_nodes_req(request, &query_opts);
+        let (mut response_body, index) = self
+            .execute_request(req, hyper::Body::empty(), query_opts.timeout)
             .await?;
         let bytes = response_body.copy_to_bytes(response_body.remaining());
-        Ok(serde_json::from_slice::<GetServiceNodesResponse>(&bytes)
-            .map_err(ConsulError::ResponseDeserializationFailed)?)
+        let response = serde_json::from_slice::<GetServiceNodesResponse>(&bytes)
+            .map_err(ConsulError::ResponseDeserializationFailed)?;
+        Ok(ResponseMeta { response, index })
     }
 
     /// Queries consul for a service and returns the Address:Port of all instances registered for that service.
     pub async fn get_service_addresses_and_ports(
         &self,
         service_name: &str,
+        query_opts: Option<QueryOptions>,
     ) -> Result<Vec<(String, u16)>> {
         let request = GetServiceNodesRequest {
             service: service_name,
             passing: true,
             ..Default::default()
         };
-        let services = self.get_service_nodes(request).await.map_err(|e| {
+        let services = self.get_service_nodes(request, query_opts).await.map_err(|e| {
             let err = format!(
                 "Unable to query consul to resolve service '{}' to a list of addresses and ports: {:?}",
                 service_name, e
@@ -534,6 +450,7 @@ impl Consul {
         })?;
 
         let addresses_and_ports = services
+            .response
             .into_iter()
             .map(Self::parse_host_port_from_service_node_response)
             .collect();
@@ -626,6 +543,7 @@ impl Consul {
     fn build_get_service_nodes_req(
         &self,
         request: GetServiceNodesRequest<'_>,
+        query_opts: &QueryOptions,
     ) -> http::request::Builder {
         let req = hyper::Request::builder().method(Method::GET);
         let mut url = String::new();
@@ -640,12 +558,7 @@ impl Consul {
         if let Some(filter) = request.filter {
             url.push_str(&format!("&filter={}", filter));
         }
-        if let Some(ns) = request.namespace {
-            url.push_str(&format!("&ns={}", ns));
-        }
-        if let Some(dc) = request.datacenter {
-            url.push_str(&format!("&dc={}", dc));
-        }
+        add_query_option_params(&mut url, query_opts, '&');
         req.uri(url)
     }
 
@@ -698,7 +611,7 @@ impl Consul {
             Ok(body) => Ok((Box::new(body), index)),
             Err(e) => {
                 span.set_status(StatusCode::Error, e.to_string());
-                return Err(ConsulError::InvalidResponse(e));
+                Err(ConsulError::InvalidResponse(e))
             }
         }
     }
@@ -728,6 +641,28 @@ impl Consul {
         }
 
         add_namespace_and_datacenter(url, request.namespace, request.datacenter)
+    }
+}
+
+fn add_query_option_params(uri: &mut String, query_opts: &QueryOptions, mut separator: char) {
+    if let Some(ns) = &query_opts.namespace {
+        if !ns.is_empty() {
+            uri.push_str(&format!("{}ns={}", separator, ns));
+            separator = '&';
+        }
+    }
+    if let Some(dc) = &query_opts.datacenter {
+        if !dc.is_empty() {
+            uri.push_str(&format!("{}dc={}", separator, dc));
+            separator = '&';
+        }
+    }
+    if let Some(idx) = query_opts.index {
+        uri.push_str(&format!("{}index={}", separator, idx));
+        separator = '&';
+        if let Some(wait) = query_opts.wait {
+            uri.push_str(&format!("{}wait={}", separator, wait.as_secs()));
+        }
     }
 }
 
@@ -780,10 +715,13 @@ mod tests {
     async fn test_register_and_retrieve_services() {
         let consul = get_client();
 
-        let new_service_name = "test-service-1".to_string();
+        let new_service_name = "test-service-44".to_string();
 
         // verify a service by this name is currently not registered
-        let service_names_before_register = consul
+        let ResponseMeta {
+            response: service_names_before_register,
+            ..
+        } = consul
             .get_all_registered_service_names(None)
             .await
             .expect("expected get_registered_service_names request to succeed");
@@ -815,7 +753,10 @@ mod tests {
             .expect("expected register_entity request to succeed");
 
         // verify the newly registered service is retrieved
-        let service_names_after_register = consul
+        let ResponseMeta {
+            response: service_names_after_register,
+            ..
+        } = consul
             .get_all_registered_service_names(None)
             .await
             .expect("expected get_registered_service_names request to succeed");
@@ -830,8 +771,8 @@ mod tests {
             passing: true,
             ..Default::default()
         };
-        let res = consul.get_service_nodes(req).await.unwrap();
-        assert_eq!(res.len(), 0);
+        let ResponseMeta { response, .. } = consul.get_service_nodes(req, None).await.unwrap();
+        assert_eq!(response.len(), 0);
 
         // TODO: Make these tests pass for nomad and crdb.
         // let req = GetServiceNodesRequest {
@@ -891,7 +832,7 @@ mod tests {
         let string_value = "This is a lock test";
         let new_string_value = "This is a changed lock test";
         let req = LockRequest {
-            key: key,
+            key,
             behavior: LockExpirationBehavior::Release,
             lock_delay: std::time::Duration::from_secs(1),
             ..Default::default()
@@ -926,7 +867,7 @@ mod tests {
         verify_single_value_matches(key_resp, &new_string_value);
 
         let req = LockRequest {
-            key: key,
+            key,
             behavior: LockExpirationBehavior::Delete,
             lock_delay: std::time::Duration::from_secs(1),
             session_id: &session_id,
@@ -944,7 +885,7 @@ mod tests {
         let key = "test/consul/watchedlock";
         let string_value = "This is a lock test";
         let req = LockRequest {
-            key: key,
+            key,
             behavior: LockExpirationBehavior::Release,
             lock_delay: std::time::Duration::from_secs(0),
             ..Default::default()
@@ -970,7 +911,7 @@ mod tests {
 
         assert!(start_index > 0);
         let watch_req = LockWatchRequest {
-            key: key,
+            key,
             consistency: ConsistencyMode::Consistent,
             index: Some(start_index),
             wait: Duration::from_secs(60),
@@ -1020,7 +961,7 @@ mod tests {
 
         let sn = ServiceNode {
             node: node.clone(),
-            service: empty_service.clone(),
+            service: empty_service,
         };
 
         let (host, port) = Consul::parse_host_port_from_service_node_response(sn);
@@ -1039,7 +980,7 @@ mod tests {
         value: &str,
     ) -> Result<(bool, u64)> {
         let req = CreateOrUpdateKeyRequest {
-            key: key,
+            key,
             ..Default::default()
         };
         Ok(consul
@@ -1049,7 +990,7 @@ mod tests {
 
     async fn read_key(consul: &Consul, key: &str) -> Result<Vec<ReadKeyResponse>> {
         let req = ReadKeyRequest {
-            key: key,
+            key,
             ..Default::default()
         };
         consul.read_key(req).await
@@ -1057,7 +998,7 @@ mod tests {
 
     async fn delete_key(consul: &Consul, key: &str) -> Result<bool> {
         let req = DeleteKeyRequest {
-            key: key,
+            key,
             ..Default::default()
         };
         consul.delete_key(req).await